move logic to the use case
